+++ app/src/main/java/me/portailler/florian/testanimation/ui/compose/sharedelements/lib/PathMotion.kt
@@ -1,0 +1,12 @@
+package me.portailler.florian.testanimation.ui.compose.sharedelements.lib
+
+import androidx.compose.ui.geometry.Offset
+import androidx.compose.ui.geometry.lerp
+
+typealias PathMotion = (start: Offset, end: Offset, fraction: Float) -> Offset
+
+typealias PathMotionFactory = () -> PathMotion
+
+val LinearMotion: PathMotion = ::lerp
+
+val LinearMotionFactory: PathMotionFactory = { LinearMotion }